Issues from comments are fixed
