+++ .gitignore
@@ -1,4 +1,4 @@
 out/*
 target/*
 *.idea
-lesson4.iml
+*.iml
